requirements.txt
appdirs==1.4.4
attrs==19.3.0
beautifulsoup4==4.9.1
cached-property==1.5.1
cachetools==4.1.1
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cssselect2==0.3.0
defusedxml==0.6.0
google-api-core==1.21.0
google-auth==1.18.0
google-cloud-core==1.3.0
google-cloud-translate==2.0.1
googleapis-common-protos==1.52.0
grpcio==1.30.0
html5lib==1.1
idna==2.10
isodate==0.6.0
Jinja2==2.11.2
lxml==4.5.1
MarkupSafe==1.1.1
ordereddict==1.1
Pillow==7.1.2
pprint==0.1
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.19.1
pycountry==19.8.18
pycparser==2.20
pycups==2.0.1
PyGObject==3.36.1
PyPDF2==1.26.0
Pyphen==0.9.5
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-toolbelt==0.9.1
rsa==4.6
six==1.15.0
soupsieve==2.0.1
tinycss2==1.0.2
urllib3==1.25.9
WeasyPrint==51
webencodings==0.5.1
WooCommerce==2.1.1
zeep==3.4.0
